test_system/requirements.txt
# Các thư viện cần thiết cho môi trường test
mysql-connector-python==8.0.32
pandas==1.5.3
numpy==1.24.3
requests==2.28.2
python-dotenv==1.0.0
pytest==7.3.1
pytest-html==3.2.0
pytest-cov==4.1.0